Small changes to separate test and teleop modes
Created separate functions for binding command triggers to be used
in test and teleop modes. The goal is to make it easier separate
controls that we'll use when playing the game from controls that'll
help us test specific motors. I was worried that if we keep them
mixed, we'll start breaking game controls in order to do tests, or
damage the robot when using game controls for testing.